backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic>=2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.7
google-generativeai==0.3.2
groq==0.4.1
mistralai==0.0.12
httpx==0.25.2
cryptography==41.0.7
pymongo==4.6.0